+++ src/main/java/io/github/betterclient/ai/training/SmallNetworkTrainer.java
@@ -4,6 +4,7 @@ import io.github.betterclient.ai.neural.NeuralLayer;
 import io.github.betterclient.ai.neural.NeuralNetwork;
 import io.github.betterclient.ai.neural.Neuron;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,23 +12,23 @@ import java.util.List;
  * More accurate training for smaller models
  */
 public class SmallNetworkTrainer {
-    public static void train(NeuralNetwork network, List<TrainingInput> samples, float h, float learningRate) {
+    public static void train(NeuralNetwork network, List<TrainingInput> samples, double h, double learningRate) {
         List<NeuralLayer> layers = new ArrayList<>(network.layers);
         layers.removeFirst();
         for (NeuralLayer neuralLayer : layers) {
-            float originalCost = network.cost(samples);
+            double originalCost = network.cost(samples);
             learn(neuralLayer, network, samples, h, originalCost, learningRate);
         }
     }
 
-    public static void learn(NeuralLayer layer, NeuralNetwork network, List<TrainingInput> samples, float h, float currentCost, float learningRate) {
+    public static void learn(NeuralLayer layer, NeuralNetwork network, List<TrainingInput> samples, double h, double currentCost, double learningRate) {
         for (Neuron neuronIn : layer.before.neurons) {
             for (Neuron neuronOut : layer.neurons) {
-                float value = neuronIn.connectionWeights.get(neuronOut);
-                neuronIn.connectionWeights.put(neuronOut, value + h);
+                double value = neuronIn.connectionWeights.get(neuronOut).doubleValue();
+                neuronIn.connectionWeights.put(neuronOut, BigDecimal.valueOf(value + h));
 
-                float cost = network.cost(samples);
-                float delta = cost - currentCost;
+                double cost = network.cost(samples);
+                double delta = cost - currentCost;
 
                 if (delta > 0) {
                     //Wrong direction
@@ -37,7 +38,7 @@ public class SmallNetworkTrainer {
                     value += learningRate;
                 }
 
-                neuronIn.connectionWeights.put(neuronOut, value);
+                neuronIn.connectionWeights.put(neuronOut, BigDecimal.valueOf(value));
 
                 currentCost = network.cost(samples);
             }
@@ -47,8 +48,8 @@ public class SmallNetworkTrainer {
 
         for (Neuron neuron : layer.neurons) {
             neuron.bias += h;
-            float currentCost0 = network.cost(samples);
-            float delta = (currentCost0 - currentCost);
+            double currentCost0 = network.cost(samples);
+            double delta = (currentCost0 - currentCost);
             if (delta != 0) {
                 //Change happened
